# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: SHIELD API
  version: 1.0.0
  description: Smart Housing Infrastructure and Entry Log Digitalization API
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /ledger-entries:
    get:
      tags: [Accounting]
      summary: List ledger entries
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Ledger entries fetched}
    post:
      tags: [Accounting]
      summary: Create ledger entry
      responses:
        '200': {description: Ledger entry created}
  /ledger-entries/{id}:
    get:
      tags: [Accounting]
      summary: Get ledger entry by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Ledger entry fetched}
    put:
      tags: [Accounting]
      summary: Update ledger entry
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Ledger entry updated}
    delete:
      tags: [Accounting]
      summary: Delete ledger entry
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Ledger entry deleted}
  /ledger-entries/account/{accountHeadId}:
    get:
      tags: [Accounting]
      summary: List ledger entries by account head
      parameters:
        - in: path
          name: accountHeadId
          required: true
          schema: {type: string, format: uuid}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Ledger entries by account head fetched}
  /ledger-entries/fund/{fundCategoryId}:
    get:
      tags: [Accounting]
      summary: List ledger entries by fund category
      parameters:
        - in: path
          name: fundCategoryId
          required: true
          schema: {type: string, format: uuid}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Ledger entries by fund category fetched}
  /ledger-entries/date-range:
    get:
      tags: [Accounting]
      summary: List ledger entries by date range
      parameters:
        - in: query
          name: fromDate
          required: true
          schema: {type: string, format: date}
        - in: query
          name: toDate
          required: true
          schema: {type: string, format: date}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Ledger entries by date range fetched}
...
  /asset-categories:
    post:
      tags: [Asset]
      summary: Create asset category
      responses:
        '200': {description: Asset category created}
    get:
      tags: [Asset]
      summary: List asset categories
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Asset categories fetched}
  /asset-categories/{id}:
    get:
      tags: [Asset]
      summary: Get asset category
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Asset category fetched}
    put:
      tags: [Asset]
      summary: Update asset category
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Asset category updated}
    delete:
      tags: [Asset]
      summary: Delete asset category
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Asset category deleted}
  /assets:
    post:
      tags: [Asset]
      summary: Create asset
      responses:
        '200': {description: Asset created}
    get:
      tags: [Asset]
      summary: List assets
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Assets fetched}
  /assets/{id}:
    get:
      tags: [Asset]
      summary: Get asset by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Asset fetched}
    put:
      tags: [Asset]
      summary: Update asset
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Asset updated}
    delete:
      tags: [Asset]
      summary: Delete asset
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Asset deleted}
  /assets/category/{categoryId}:
    get:
      tags: [Asset]
      summary: List assets by category
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: {type: string, format: uuid}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Assets by category fetched}
  /assets/location/{location}:
    get:
      tags: [Asset]
      summary: List assets by location
      parameters:
        - in: path
          name: location
          required: true
          schema: {type: string}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Assets by location fetched}
  /assets/tag/{tag}:
    get:
      tags: [Asset]
      summary: Get asset by tag
      parameters:
        - in: path
          name: tag
          required: true
          schema: {type: string}
      responses:
        '200': {description: Asset fetched by tag}
  /assets/verify-qr/{qrCode}:
    get:
      tags: [Asset]
      summary: Verify asset by QR code
      parameters:
        - in: path
          name: qrCode
          required: true
          schema: {type: string}
      responses:
        '200': {description: Asset verified}
  /assets/amc-expiring:
    get:
      tags: [Asset]
      summary: List assets with AMC expiry
      parameters:
        - in: query
          name: days
          required: false
          schema: {type: integer, default: 30}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: AMC expiring assets fetched}
  /assets/warranty-expiring:
    get:
      tags: [Asset]
      summary: List assets with warranty expiry
      parameters:
        - in: query
          name: days
          required: false
          schema: {type: integer, default: 30}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Warranty expiring assets fetched}
...
  /complaints/asset/{assetId}:
    get:
      tags: [Complaint]
      summary: List complaints by asset
      parameters:
        - in: path
          name: assetId
          required: true
          schema: {type: string, format: uuid}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Complaints by asset fetched}
...
  /preventive-maintenance/asset/{assetId}:
    get:
      tags: [Asset]
      summary: List preventive maintenance by asset
      parameters:
        - in: path
          name: assetId
          required: true
          schema: {type: string, format: uuid}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Preventive maintenance by asset fetched}
...
  /asset-depreciation/asset/{assetId}:
    get:
      tags: [Asset]
      summary: List asset depreciation by asset
      parameters:
        - in: path
          name: assetId
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Asset depreciation by asset fetched}
...
  /documents/category/{categoryId}:
    get:
      tags: [Documents]
      summary: List documents by category
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Category documents fetched}
...
  /marketplace-listings/category/{categoryId}:
    get:
      tags: [Marketplace]
      summary: List marketplace listings by category
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Category marketplace listings fetched}
...
  /config/category/{category}:
    get:
      tags: [Configuration]
      summary: List tenant configuration by category
      parameters:
        - in: path
          name: category
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Tenant configuration by category fetched}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Sort:
      in: query
      name: sort
      schema:
        type: string
        example: createdAt,desc
...